--- a/scripts/docs/social-elements-test-report.md
+++ b/scripts/docs/social-elements-test-report.md
@@ -1,0 +1,172 @@
+# Social Elements Test Report
+
+## Executive Summary
+
+We've conducted comprehensive testing of both dub.co link shortening and social card generation systems. Here are the key findings:
+
+## 🎯 Test Results Overview
+
+### ✅ Working Components
+- **Social Card Generation**: ✅ **WORKING** - Fly.io service is operational
+- **Social Card Speed**: ✅ **GOOD** - Average 579ms response time
+- **Cache Headers**: ✅ **WORKING** - Proper CDN caching implemented
+- **Meta Tags**: ✅ **WORKING** - Proper Open Graph metadata
+- **Fallback Data**: ✅ **WORKING** - Service works without database
+
+### ⚠️ Issues Identified
+- **Dub.co API**: ⚠️ **API KEY MISSING** - VITE_DUB_CO_KEY not found in development
+- **Response Times**: ⚠️ **COULD BE FASTER** - 579ms average (target: <300ms)
+
+---
+
+## 📊 Social Card Performance Analysis
+
+### Speed Test Results (8 different card types tested)
+
+| Card Type | Response Time | Status | Data Source |
+|-----------|---------------|--------|-------------|
+| Next.js Repo | 497ms | ✅ | Fallback |
+| Angular Repo | 550ms | ✅ | Fallback |
+| User Card (tj) | 555ms | ✅ | Fallback |
+| React Repo | 557ms | ✅ | Fallback |
+| Vue Repo | 558ms | ✅ | Fallback |
+| User Card (gaearon) | 558ms | ✅ | Fallback |
+| Custom Title | 642ms | ✅ | Database |
+| Home Card | 712ms | ✅ | Database |
+
+### Performance Insights
+
+- **100% Success Rate**: All cards generated successfully
+- **Cache Headers**: Proper `max-age=3600, s-maxage=86400` (1h client, 24h CDN)
+- **Server Processing**: 31-234ms server-side generation time
+- **Data Sources**: 
+  - Fallback data: Faster (497-558ms)
+  - Database data: Slower (642-712ms)
+
+---
+
+## 🔗 Dub.co Link Shortening Analysis
+
+### Current Status
+- **API Key**: ❌ Missing `VITE_DUB_CO_KEY` in development environment
+- **Development Mode**: ✅ Properly falls back to original URLs
+- **Production Setup**: ⚠️ Domain configured for `oss.fyi` vs `dub.sh`
+- **Code Quality**: ✅ Proper error handling and fallbacks
+
+### Code Analysis (`src/lib/dub.ts`)
+```javascript
+// In development, returns original URL (expected behavior)
+if (isDev) {
+  return { shortLink: url }; // Original URL returned
+}
+
+// In production, would use oss.fyi domain
+const DOMAIN = isDev ? "dub.sh" : "oss.fyi";
+```
+
+---
+
+## 🏗️ Architecture Review
+
+### Social Cards System
+```
+Frontend → Fly.io Service → Supabase (optional) → SVG Generation
+            (579ms avg)      (fallback available)
+```
+
+**Strengths:**
+- Redundant fallback data system
+- Proper CDN caching
+- Fast SVG generation
+- Rate limiting (60 req/min)
+
+**Optimization Opportunities:**
+- Pre-generate popular repository cards
+- Implement edge caching
+- Database query optimization
+
+### Dub.co Integration
+```
+Frontend → Dub.co API → Short URL Response
+           (if api key)    (with analytics)
+```
+
+**Strengths:**
+- Graceful degradation without API key
+- UTM parameter tracking
+- Custom domain support (oss.fyi)
+
+**Issues:**
+- Missing API key in development
+- No production testing capability
+
+---
+
+## 🛠️ Recommendations
+
+### Immediate Actions (High Priority)
+
+1. **Configure Dub.co API Key**
+   ```bash
+   # Add to Netlify environment
+   netlify env:set VITE_DUB_CO_KEY "dub_xxxxx"
+   ```
+
+2. **Test Production Dub.co Integration**
+   - Verify API key works in production
+   - Test `oss.fyi` domain functionality
+   - Validate URL shortening flow
+
+### Performance Optimizations (Medium Priority)
+
+3. **Improve Social Card Speed**
+   - Target: <300ms average response time
+   - Pre-cache popular repositories
+   - Optimize database queries
+   - Consider edge deployment
+
+4. **Enhanced Monitoring**
+   - Add performance metrics to dashboard
+   - Monitor success rates
+   - Track cache hit ratios
+
+### Nice-to-Have Improvements (Low Priority)
+
+5. **Advanced Features**
+   - A/B test different card designs
+   - Dynamic social card sizing
+   - Real-time analytics dashboard
+
+---
+
+## 🧪 Test Files Created
+
+1. **`test-social-elements.html`** - Interactive browser test suite
+2. **`test-social-card-speed.js`** - Comprehensive speed testing
+3. **`test-dub-api-direct.js`** - Direct API testing script
+
+## 🚀 Quick Validation Commands
+
+```bash
+# Test social card speed
+node test-social-card-speed.js
+
+# Test dub.co with API key
+VITE_DUB_CO_KEY="your_key" node test-dub-api-direct.js
+
+# Open interactive test in browser
+open http://localhost:5174/test-social-elements.html
+```
+
+---
+
+## 🎉 Conclusion
+
+The social elements system is **mostly functional** with good performance characteristics:
+
+- **Social Cards**: ✅ Working well, good speed, 100% reliability
+- **Dub.co Integration**: ⚠️ Needs API key configuration but code is solid
+
+**Overall Grade: B+ (Very Good)**
+
+Main blocker is the missing Dub.co API key, which is a configuration issue rather than a code problem. Once that's resolved, the system should work excellently.