… scripts/testing-tools/README.md
… scripts/testing-tools/test-social-elements.html
… tasks/prd-social-card-speed-optimization.md